pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ehio"
version = "0.1.0"
description = "Bridge between Airtable metadata and Drakkar workflows, with SFTP output transfer."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pyairtable>=2.3",
  "paramiko>=3.4",
  "pyyaml>=6",
]

[project.optional-dependencies]
dev = ["pytest>=8"]
release = [
  "build>=1.2",
  "twine>=5",
]

[project.scripts]
ehio = "ehio.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
ehio = ["data/*.yaml"]
